EOS-20956-NWportHealthView: Added common error class, other minor changes.
